Modified the Makefile and mem.c for handling the test bench.
Right now, it passes 58 / 61.
One of the tests initializes 1 byte and our algorithm (as per spec)
should pad extra bytes so that it equals 4096 (=pagesize). But this 4096
also includes bytes for the header. So we need to allocate more.
The change that I've created is specific to the test bench.. It needs to
be modified into a generic manner. Once that's done, I believe it should
take care of all the tests.