Implemented the Extra Credit but doesn't pass tests
Got the file size from stat. But the server kept hanging for no reason.
Tested with gdb and wget but no clue. Then the TA asked to modify the
implementation. Modified it, and it passes 1 of the EC once, but it was
not repetitive. Need to look more into requestHandle().